[fix-polygon-reflex-angles] fix stroke width
